sqaud working

+++ CricketAppClient/src/app/_admin/players/players.component.ts
@@ -1,8 +1,13 @@
-import { Component, EventEmitter, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
-import { Pagination } from 'src/app/models/pagination';
-import { players } from 'src/app/models/players';
+import { Component, EventEmitter, OnInit, ViewChild } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { ModalDirective } from 'ngx-bootstrap/modal';
+import { ToastrService } from 'ngx-toastr';
+import { Pagination } from 'src/app/_models/pagination';
+import { players } from 'src/app/_models/players';
+import { AppcommonService } from 'src/app/_services/appcommon.service';
+import { ConfirmService } from 'src/app/_services/confirm.service';
 import { PlayersService } from 'src/app/_services/players.service';
+import { AddplayerComponent } from 'src/app/_ui_component/addplayer/addplayer.component';
 
 @Component({
   selector: 'app-players',
@@ -20,21 +25,29 @@ export class PlayersComponent implements OnInit {
   player: EventEmitter<players> = new EventEmitter();
   buttonText:string="Add";
 
+  @ViewChild('playerModal', { static: false }) modal?: ModalDirective;
 
-  constructor(private fb: FormBuilder,private playerService:PlayersService) { }
+  @ViewChild(AddplayerComponent,{static:false}) child!: AddplayerComponent ;
+
+
+  constructor(private fb: FormBuilder,private playerService:PlayersService,private confirmService:ConfirmService,
+    private appcommonService:AppcommonService,private toastr:ToastrService) { }
 
   ngOnInit(): void {
-    this.intitializeForm();
     this.loadPlayers();
   }
-  intitializeForm()
+  onShow()
   {
-    this.playerForm=this.fb.group({
-      Name: [''],
-      IsLocalPlayer: [''],
-      InternationalTeam:[''],
-      IsActive:[false]
-    });
+    this.modal?.show();
+  }
+  onClose()
+  {
+     this.child.OnReset();
+     this.modal?.hide();
+  }
+  getappcommonService(indx:number)
+  {
+    return this.appcommonService.generateRandom(indx);
   }
   loadPlayers() {
     this.playerService.getPlayers(this.pageNumber,this.pageSize).subscribe(response => {
@@ -42,9 +55,17 @@ export class PlayersComponent implements OnInit {
       this.pagination = response.pagination;
     })
   }
-  outputFunction(event:any)
+  outputFunction(event:string)
   {
-    this.loadPlayers();
+    if(event=="update"||event=="add"||event=="delete")
+    {
+      this.loadPlayers();
+      this.onClose();
+    }
+    else if(event=="reset")
+    {
+      // this.onClose();
+    }
   }
   pageChanged(event: any)
   {
@@ -54,13 +75,21 @@ export class PlayersComponent implements OnInit {
  
   OnEdit(item:any)
   {
+    this.onShow();
     this.player.emit(item);
   }
   OnDelete(id:any)
   {
-    this.playerService.deletePlayer(id).subscribe(response => {
-      this.loadPlayers();
-    });
+    this.confirmService.confirm('Confirm delete').subscribe(result => {
+      if (result) {
+        this.playerService.deletePlayer(id).subscribe(response => {
+          if(response==true)
+          {
+            this.toastr.success('Remove Player', 'Removed successfully !');
+            this.loadPlayers();
+          }
+        });
+      }
+    })
   }
-
 }
